refactor: separate header component
