+++ src/components/header/container/header.js
@@ -1,9 +1,9 @@
 import { connect } from 'react-redux';
-import TodoList from '../component/todoList';
+import Header from '../component/header';
 
 const mapStateToProps = (state) => ({
   name: state.name,
   count: state.tasks.data.filter((i) => i.isActive === true).length,
 });
 
-export default connect(mapStateToProps)(TodoList);
+export default connect(mapStateToProps)(Header);
