+++ src/components/home/reducer.js
@@ -1,4 +1,4 @@
 export default function reducer(state = 'anonymous', action) {
-  if (action.type === 'ENTER_NAME') return action.text;
+  if (action.type === 'ENTER_NAME') return action.text ? action.text : 'anonymous';
   return state;
 }
